[NS-72] Re-organize plan code
- Move all plan classes over to `io.sparkcyclone.spark.plans` package

- Fold plan-relevant code from `io.sparkcyclone.spark.planning` over to
`io.sparkcyclone.spark.plans` package
