+++ src/main/scala/io/sparkcyclone/spark/planning/VeFunction.scala
@@ -3,6 +3,7 @@ package io.sparkcyclone.spark.planning
 import io.sparkcyclone.native.NativeFunction
 import io.sparkcyclone.spark.agile.core.CodeLines
 import io.sparkcyclone.spark.agile.core.{CVector, VeType}
+import io.sparkcyclone.spark.plans.PlanCallsVeFunction
 import io.sparkcyclone.spark.planning.LibLocation.LibLocation
 import java.nio.file.Path
