+++ src/main/scala/io/sparkcyclone/spark/plans/VeAmplifyBatchesPlan.scala
@@ -1,9 +1,9 @@
-package io.sparkcyclone.spark.planning.plans
+package io.sparkcyclone.spark.plans
 
 import io.sparkcyclone.cache.ArrowEncodingSettings
 import io.sparkcyclone.data.vector.{VeBatchOfBatches, VeColBatch}
 import io.sparkcyclone.plugin.SparkCycloneExecutorPlugin.{source, veProcess, vectorEngine}
-import io.sparkcyclone.spark.planning.{PlanCallsVeFunction, PlanMetrics, SupportsVeColBatch, VeFunction}
+import io.sparkcyclone.spark.planning.VeFunction
 import io.sparkcyclone.util.CallContext
 import io.sparkcyclone.util.CallContextOps._
 import com.typesafe.scalalogging.LazyLogging
