+++ src/main/scala/io/sparkcyclone/spark/plans/VeHashExchangePlan.scala
@@ -1,4 +1,4 @@
-package io.sparkcyclone.spark.planning.plans
+package io.sparkcyclone.spark.plans
 
 import io.sparkcyclone.data.vector.VeColBatch
 import io.sparkcyclone.plugin.SparkCycloneExecutorPlugin.{source, veProcess, vectorEngine}
