+++ src/main/scala/io/sparkcyclone/spark/rules/CombinedCompilationColumnarRule.scala
@@ -4,9 +4,9 @@ import io.sparkcyclone.native.{CompiledCodeInfo, NativeFunction}
 import io.sparkcyclone.plugin.SparkCycloneDriverPlugin
 import io.sparkcyclone.spark.agile.core.CodeLines
 import io.sparkcyclone.spark.agile.CodeStructure
+import io.sparkcyclone.spark.plans.PlanCallsVeFunction.UncompiledPlan
 import io.sparkcyclone.spark.planning._
 import io.sparkcyclone.spark.planning.LibLocation.DistributedLibLocation
-import io.sparkcyclone.spark.planning.PlanCallsVeFunction.UncompiledPlan
 import io.sparkcyclone.spark.planning.VeFunctionStatus._
 import java.nio.file.Path
 import com.typesafe.scalalogging.LazyLogging
